predict.LogitBoost {caTools}	R Documentation

Prediction Based on LogitBoost Classification Algorithm

Description

Prediction or Testing using logitboost classification algorithm

Usage

predict.LogitBoost(object, xtest, type = c("class", "raw"), nIter=NA, ...)

Arguments

object 	 An object of class "LogitBoost" see "Value" section of LogitBoost for details

xtest 	 A matrix or data frame with test data. Rows contain samples and columns contain features

type 	 See "Value" section

nIter 	 An optional integer, used to lower number of iterations (decision stumps) used in the decision making. If not provided than the number will be the same as the one provided in LogitBoost. If provided than the results will be the same as running LogitBoost with fewer iterations.

... 	 not used but needed for compatibility with generic predict method

Details

Logitboost algorithm relies on a voting scheme to make classifications. Many (nIter of them) week classifiers are applied to each sample and their findings are used as votes to make the final classification. The class with the most votes "wins". However, with this scheme it is common for two cases have a tie (the same number of votes), especially if number of iterations is even. In that case NA is returned, instead of a label.

Value

If type = "class" (default) label of the class with maximal probability is returned for each sample. If type = "raw", the a-posterior probabilities for each class are returned.

Author(s)

Jarek Tuszynski (SAIC) jaroslaw.w.tuszynski@saic.com

See Also

LogitBoost has training half of LogitBoost code

Examples

# See LogitBoost example

[Package caTools version 1.9 Index]